requirements.txt
mlx-lm
mlx-whisper
cmudict
python-dotenv
fastapi
uvicorn
pipecat-ai[silero,webrtc,daily,openai,google,cartesia]
pipecatcloud